added basic frontend

+++ fastapi-backend/app/models/payment.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, String, DateTime
+from sqlalchemy import Column, Integer, Float, ForeignKey, String, DateTime
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from app.database.connection import Base
@@ -8,7 +8,7 @@ class Payment(Base):
     
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("users.id"))
-    amount = Column(Integer, nullable=False)
+    amount = Column(Float, nullable=False)
     status = Column(String(50), default="pending")  # pending, completed, failed
     timestamp = Column(DateTime, default=datetime.utcnow)
